Move protocols out of `dag` package
